rename util/__init__.py to utils.py
