+++ imjoy/runners/subprocess.py
@@ -15,7 +15,7 @@ import GPUtil
 from imjoy import __version__, API_VERSION, HERE
 from imjoy.connection.decorator import socketio_handler as sio_on
 from imjoy.connection.server import NAME_SPACE
-from imjoy.util import console_to_str, parse_repos, get_psutil, kill_process
+from imjoy.utils import console_to_str, parse_repos, get_psutil, kill_process
 from .helper import (
     apply_conda_activate,
     install_reqs,
