+++ imjoy/services/terminal.py
@@ -10,7 +10,7 @@ import sys
 from imjoy import __version__
 from imjoy.connection.decorator import socketio_handler as sio_on
 from imjoy.connection.server import NAME_SPACE
-from imjoy.util import get_psutil
+from imjoy.utils import get_psutil
 
 if sys.platform != "win32":
     import fcntl
